api/requirements.txt
fastapi
uvicorn
sentence-transformers
qdrant-client
google-generativeai

